requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
Jinja2==3.1.5
python-dotenv==1.0.1
aiofiles==24.1.0
aio-pika==9.5.4
motor==3.6.0